% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_generate_blurb.R
\name{generate_blurb}
\alias{generate_blurb}
\title{Generate Blurb Statistics}
\usage{
generate_blurb(df)
}
\arguments{
\item{df}{A data frame containing population data}
}
\value{
a list containing total population, percentage of population that is
male, percentage of population that is female, total population of individuals
below 5 years of age, and total population that is in the working class
}
\description{
Given a data frame of data (df), generates total population,
percentage of males, percentage of females, total population below 5 years of
age, and total working class people (age 15-65)
}
\examples{
generate_blurb(country_data)

}
