% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rename_files.R
\name{rename_file}
\alias{rename_file}
\title{Rename a file}
\usage{
rename_file(file, new_name)
}
\arguments{
\item{file}{The file that you want to rename}

\item{new_name}{The name that you want to rename the original file to}
}
\value{
the new name of the file
}
\description{
Given a file (file) and what you want to rename it
to (new_name), renames a file
}
\examples{
rename_file("country1.pdf","country2")


}
