% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_all.R
\name{render_all}
\alias{render_all}
\title{Batch Render Reports for an ADM Level}
\usage{
render_all(df, adm_level, country_name, year)
}
\arguments{
\item{df}{A data frame containing population data}

\item{adm_level}{A numeric specifying ADM level}

\item{country_name}{A string of the name of the country}

\item{year}{A numeric specifying year}
}
\value{
renders all reports report stored in a folder in the /reports/
directory
}
\description{
Given a data frame of data (df), a specified administrative
level (adm_level), the name of the country (country_name), and a specified
year (year), generates a batch of reports for every unique entity in the
given administrative level
}
\examples{
render_all(country_data,1,"USA",2017 )


}
\author{
Kevin Gui
}
