% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_quarto_doc.R
\name{render_quarto_document}
\alias{render_quarto_document}
\title{Render Quarto Document}
\usage{
render_quarto_document(
  df = NULL,
  shp = NULL,
  adm_level = NULL,
  adm_name = NULL,
  country_name = NULL,
  year = NULL,
  parameterized = TRUE,
  batch = FALSE
)
}
\arguments{
\item{df}{A data frame provided}

\item{shp}{A shape file}

\item{adm_level}{A numeric specifying ADM level}

\item{adm_name}{A string specifying the name of the ADM}

\item{country_name}{A string of the name of the country}

\item{year}{A numeric specifying year}

\item{parameterized}{Whether asking for a parameterized report or not}

\item{batch}{Whether asking for a batch report or not}
}
\description{
Render the Quarto document with the given parameters.
}
