% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_single.R
\name{render_single}
\alias{render_single}
\title{Render a Single Parameterized Report}
\usage{
render_single(df, shp, adm_level, adm_name, country_name, year)
}
\arguments{
\item{df}{A data frame provided}

\item{shp}{A shape file}

\item{adm_level}{A numeric specifying ADM level}

\item{adm_name}{A string specifying the name of the ADM}

\item{country_name}{A string of the name of the country}

\item{year}{A numeric specifying year}
}
\value{
renders a single report
}
\description{
Given a data frame containing population data (df), a specified
administrative level (adm_level), the name of the region (adm_name), the
name of the country (country_name), and the year (yr), return a single
report based on the given parameters
}
\examples{
render_single(country_data,1,adm_1_shape,"California","USA",2017)


}
\author{
Kevin Gui
}
