% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validate_paramters.R
\name{validate_parameters}
\alias{validate_parameters}
\title{Validate Parameters}
\usage{
validate_parameters(
  adm_level = 0,
  adm_name = NULL,
  df = NULL,
  year = NULL,
  parameterized = TRUE
)
}
\arguments{
\item{adm_level}{A numeric specifying ADM level}

\item{adm_name}{A string specifying the name of the ADM}

\item{df}{A data frame provided}

\item{year}{A numeric specifying year}

\item{parameterized}{Whether parameterized report or not}
}
\description{
Validate the provided parameters to ensure they are correct.
}
