% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_national.R
\name{render_national}
\alias{render_national}
\title{Render a National Report}
\usage{
render_national(df, country_name, year)
}
\arguments{
\item{df}{A data frame provided}

\item{country_name}{A string of the name of the country}

\item{year}{A numeric specifying year}
}
\value{
renders a single national report stored in the /reports/ directory
}
\description{
Given a data frame containing population data (df), the
name of the country (country_name), and the year (yr), return a single
national report based on the given parameters
}
\examples{
render_national(country_data,"USA",2017)


}
\author{
Kevin Gui
}
